+++ hvega/CHANGELOG.md
@@ -1,6 +1,42 @@
 For the latest version of this document, please see
 [https://github.com/DougBurke/hvega/blob/master/hvega/CHANGELOG.md](https://github.com/DougBurke/hvega/blob/master/hvega/CHANGELOG.md).
 
+## 0.6.0.0
+
+The Vega-Lite tests are now validated against version 4.2.0 of the
+Vega-Lite schema.
+
+### New functionality
+
+New function for use with `encoding`: `strokeDash`.
+
+### Breaking changes
+
+The constructors for `FacetConfig` have been renamed from `FColumns`
+and `FSpacing` to `FacetColumns` and `FacetSpacing`. This is to
+support the new `FSpacing` constructor for `FacetChannel`.
+
+### New constructors
+
+`FacetChannel` has gained the following constructors: `FAlign`,
+`FCenter`, and `FSpacing`. The last one has caused the renaming of the
+constructors for the `FacetConfig` type.
+
+`MSymbol` has been added to `MarkChannel` which can be used to make the
+`shape` encoding conditional on a data or selection condition.
+
+The `TUStep` and `TUMaxBins` constructors have been added to `TimeUnit`
+for controlling how time values are binned.
+
+The `MarkProperty` type has gained the `MCornerRadiusEnd` constructor,
+which is used to draw rounded histogram bars.
+
+The `ScaleProperty` type has gained `SDomainMid`, useful for asymmetric
+diverging color scales.
+
+Labels can now be vertically aligned to their baseline with the
+`AlignBaseline` constructor of the `VAlign` type.
+
 ## 0.5.0.0
 
 Update to version 4.0 of the Vega-Lite specification (tested against